Fix loading screen page padding
